requirements.txt
numpy
pillow
flask
# for anaconda use first # for google direct link for linux
#torch==1.10.0+cu102
https://download.pytorch.org/whl/cu102/torch-1.10.0%2Bcu102-cp39-cp39-linux_x86_64.whl
#https://download.pytorch.org/whl/cu102/torch-1.10.0%2Bcu102-cp39-cp39-win_amd64.whl

#torchvision==0.11.1+cu102 -f https://download.pytorch.org/whl/cu102/torch_stable.html
https://download.pytorch.org/whl/cu102/torchvision-0.11.1%2Bcu102-cp39-cp39-linux_x86_64.whl
#https://download.pytorch.org/whl/cu102/torchvision-0.11.1%2Bcu102-cp39-cp39-win_amd64.whl